Fix: Don't truncate rpm value. Add gauge name label.

+++ src/hud.h
@@ -33,7 +33,7 @@ public:
 		FONT & lcdfont, FONT & sansfont, float displaywidth, float displayheight,
 		float curlap, float lastlap, float bestlap, float stagingtimeleft,
 		int curlapnum, int numlaps, int curplace, int numcars,
-		int rpm, int redrpm, int maxrpm,
+		float rpm, float redrpm, float maxrpm,
 		float speed, float maxspeed, bool mph, float clutch, int newgear,
 		const std::string & debug_string1, const std::string & debug_string2,
 		const std::string & debug_string3, const std::string & debug_string4,
